+++ R/beamer_presentation.R
@@ -2,40 +2,40 @@
 #'
 #' Format for converting from R Markdown to a Beamer presentation.
 #'
-#' See the \href{https://bookdown.org/yihui/rmarkdown/beamer-presentation.html}{online
-#' documentation} for additional details on using the \code{beamer_presentation}
-#' format.
+#' See the [online documentation](https://bookdown.org/yihui/rmarkdown/beamer-presentation.html)
+#' for additional details on using the `beamer_presentation` format.
 #'
 #' Creating Beamer output from R Markdown requires that LaTeX be installed.
 #'
 #' R Markdown documents can have optional metadata that is used to generate a
 #' document header that includes the title, author, and date. For more details
-#' see the documentation on R Markdown \link[=rmd_metadata]{metadata}.
+#' see the documentation on R Markdown [metadata][rmd_metadata].
 #'
 #' R Markdown documents also support citations. You can find more information on
 #' the markdown syntax for citations in the
-#' \href{https://pandoc.org/MANUAL.html#citations}{Bibliographies
-#' and Citations} article in the online documentation.
+#' [Bibliographies and Citations](https://pandoc.org/MANUAL.html#citations)
+#' article in the online documentation.
 #' @inheritParams pdf_document
 #' @inheritParams html_document
-#' @param toc \code{TRUE} to include a table of contents in the output (only
+#' @param toc `TRUE` to include a table of contents in the output (only
 #'   level 1 headers will be included in the table of contents).
 #' @param slide_level The heading level which defines individual slides. By
 #'   default this is the highest header level in the hierarchy that is followed
 #'   immediately by content, and not another header, somewhere in the document.
 #'   This default can be overridden by specifying an explicit
-#'   \code{slide_level}.
-#' @param incremental \code{TRUE} to render slide bullets incrementally. Note
+#'   `slide_level`.
+#' @param incremental `TRUE` to render slide bullets incrementally. Note
 #'   that if you want to reverse the default incremental behavior for an
-#'   individual bullet you can precede it with \code{>}. For example:
-#'   \emph{\code{> - Bullet Text}}
+#'   individual bullet you can precede it with `>`. For example:
+#'   *`> - Bullet Text`*. See more in
+#'   [Pandoc's Manual](https://pandoc.org/MANUAL.html#incremental-lists)
 #' @param theme Beamer theme (e.g. "AnnArbor").
 #' @param colortheme Beamer color theme (e.g. "dolphin").
 #' @param fonttheme Beamer font theme (e.g. "structurebold").
-#' @param self_contained Whether to generate a full LaTeX document (\code{TRUE})
-#'   or just the body of a LaTeX document (\code{FALSE}). Note the LaTeX
-#'   document is an intermediate file unless \code{keep_tex = TRUE}.
-#' @return R Markdown output format to pass to \code{\link{render}}
+#' @param self_contained Whether to generate a full LaTeX document (`TRUE`)
+#'   or just the body of a LaTeX document (`FALSE`). Note the LaTeX
+#'   document is an intermediate file unless `keep_tex = TRUE`.
+#' @return R Markdown output format to pass to [render()]
 #' @examples
 #' \dontrun{
 #'
@@ -47,6 +47,7 @@
 #' # specify an option for incremental rendering
 #' render("pres.Rmd", beamer_presentation(incremental = TRUE))
 #' }
+#' @md
 #' @export
 beamer_presentation <- function(toc = FALSE,
                                 slide_level = NULL,
@@ -106,8 +107,7 @@ beamer_presentation <- function(toc = FALSE,
     args <- c(args, pandoc_variable_arg("fonttheme", fonttheme))
 
   # highlighting
-  if (!is.null(highlight))
-    highlight <- match.arg(highlight, highlighters())
+  if (!is.null(highlight)) highlight <- resolve_highlight(highlight, highlighters())
   args <- c(args, pandoc_highlight_args(highlight))
 
   # latex engine
@@ -137,8 +137,15 @@ beamer_presentation <- function(toc = FALSE,
     # save files dir (for generating intermediates)
     saved_files_dir <<- files_dir
 
+    # make sure --include-in-header from command line will not completely
+    # override header-includes in metadata but give the latter lower precedence:
+    # https://github.com/rstudio/rmarkdown/issues/1359
+    # Same in PDF for beamer
+    # https://github.com/rstudio/rmarkdown/issues/2294
+    args <- append_in_header(process_header_includes(metadata))
+
     # no-op other than caching dir location
-    invisible(NULL)
+    args
   }
 
   # generate intermediates (required to make resources available for publish)