+++ R/pandoc.R
@@ -472,50 +472,25 @@ pandoc_self_contained_html <- function(input, output) {
 }
 
 
-validate_self_contained <- function(mathjax) {
-
-  if (identical(mathjax, "local"))
+validate_self_contained <- function(math) {
+  if (identical(math$engine, "mathjax") && identical(math$url, "local"))
     stop2("Local MathJax isn't compatible with self_contained\n",
          "(you should set self_contained to FALSE)")
 }
 
-pandoc_mathjax_args <- function(mathjax,
-                                template,
-                                self_contained,
-                                files_dir,
-                                output_dir) {
-  args <- c()
+pandoc_math_engines <- function() {
+  c("mathjax", "mathml", "webtex", "katex", "gladtex")
+}
 
-  if (!is.null(mathjax)) {
+pandoc_math_args <- function(engine = pandoc_math_engines(), url = NULL) {
 
-    if (identical(mathjax, "default")) {
-      if (identical(template, "default"))
-        mathjax <- default_mathjax()
-      else
-        mathjax <- NULL
-    }
-    else if (identical(mathjax, "local")) {
-      mathjax_path <- pandoc_mathjax_local_path()
-      mathjax_path <- render_supporting_files(mathjax_path,
-                                              files_dir,
-                                              "mathjax-local")
-      mathjax <- paste(normalized_relative_to(output_dir, mathjax_path), "/",
-                       mathjax_config(), sep = "")
-    }
+  engine <- match.arg(engine)
 
-    if (identical(template, "default")) {
-      args <- c(args, "--mathjax")
-      args <- c(args, "--variable", paste0("mathjax-url:", mathjax))
-    } else if (!self_contained) {
-      args <- c(args, paste(c("--mathjax", mathjax), collapse = "="))
-    } else {
-      warning("MathJax doesn't work with self_contained when not ",
-              "using the rmarkdown \"default\" template.", call. = FALSE)
-    }
-
+  if (!is.null(url) && engine %in% c("mathml", "gladtex")) {
+    stop2(sprintf("%s does not support setting a URL.", engine))
   }
 
-  args
+  c(paste0("--", engine), if (!is.null(url)) url)
 }
 
 
@@ -553,27 +528,49 @@ unix_mathjax_path <- function() {
 
 
 pandoc_html_highlight_args <- function(template,
-                                       highlight) {
+                                       highlight,
+                                       highlight_downlit = FALSE) {
+
+  # Reminder: we do not use pandoc_path_arg() for argument to --highlight-style
+  # https://github.com/rstudio/rmarkdown/issues/1976
 
   args <- c()
 
-  if (is.null(highlight)) {
-    args <- c(args, "--no-highlight")
-  }
-  else if (!identical(template, "default")) {
-    if (identical(highlight, "default"))
-      highlight <- "pygments"
-    args <- c(args, "--highlight-style", highlight)
-  }
-  else {
-    highlight <- match.arg(highlight, html_highlighters())
-    if (is_highlightjs(highlight)) {
-      args <- c(args, "--no-highlight")
-      args <- c(args, "--variable", "highlightjs=1")
+  # no highlighting engine
+  if (is.null(highlight)) return(pandoc_highlight_args(NULL))
+
+  highlight <- resolve_highlight(highlight, html_highlighters())
+
+  check_highlightjs <- function(highlight, engine) {
+    if (highlight != "default" && is_highlightjs(highlight)) {
+      stop(
+        sprintf(c(
+          "'%s' theme is for highlightjs highlighting engine ",
+          "and can't be used with %s engine."), c(highlight, engine)),
+        call. = FALSE
+      )
     }
-    else {
-      args <- c(args, "--highlight-style", highlight)
-    }
+  }
+
+  # downlit engine
+  if (highlight_downlit) {
+    check_highlightjs(highlight, "downlit")
+    default <- if (pandoc2.0()) resolve_highlight("arrow") else "pygments"
+    args <- c(
+      pandoc_highlight_args(highlight, default = default),
+      # variable used to insert some css in a Pandoc template
+      pandoc_variable_arg("highlight-downlit")
+    )
+  } else if (identical(template, "default") && is_highlightjs(highlight)) {
+    # highlightjs engine for default template only
+    args <- c(pandoc_highlight_args(NULL),
+              # variable used to insert some css and js
+              # in the Pandoc default template
+              pandoc_variable_arg("highlightjs", "1"))
+  } else {
+    # Pandoc engine
+    check_highlightjs(highlight, "Pandoc")
+    args <- pandoc_highlight_args(highlight, default = "pygments")
   }
 
   args
@@ -581,6 +578,38 @@ pandoc_html_highlight_args <- function(template,
 
 is_highlightjs <- function(highlight) {
   !is.null(highlight) && (highlight %in% c("default", "textmate"))
+}
+
+resolve_highlight <- function(highlight, supported = highlighters()) {
+  # for backward compatibility, partial match still need to work
+  i <- pmatch(highlight, supported, nomatch = 0L)
+  if (i > 0L) return(supported[i])
+
+  # Otherwise it could be a custom (built-in) .theme file
+  if (!pandoc2.0()) {
+    stop("Using a custom highlighting style requires Pandoc 2.0 and above",
+         call. = FALSE)
+  }
+  custom <- list(
+    # from distill
+    # https://raw.githubusercontent.com/apreshill/distill/arrow/inst/rmarkdown/templates/distill_article/resources/arrow.theme
+    arrow = pkg_file_highlight("arrow.theme"),
+    # from distill
+    # https://github.com/rstudio/distill/blob/c98d332192ff75f268ddf69bddace34e4db6d89b/inst/rmarkdown/templates/distill_article/resources/rstudio.theme
+    rstudio = pkg_file_highlight("rstudio.theme")
+  )
+  # if not an alias use the provided custom path
+  highlight <- custom[[highlight]] %||% highlight
+  # Check for extension or give informative error otherwise
+  if (!identical(xfun::file_ext(highlight), "theme")) {
+    msg <- c(
+      sprintf("`highlight` argument must be one of %s",
+              knitr::combine_words(c(supported, names(custom)), and = " or ", before = "`")),
+      " or a file with extension `.theme`."
+    )
+    stop(msg, call. = FALSE)
+  }
+  highlight
 }
 
 #' Find the \command{pandoc} executable