+++ R/render.R
@@ -834,6 +834,10 @@ render <- function(input,
 
   if (run_pandoc) {
 
+    # set env vars required during Pandoc processing
+    lua_env_vars <- xfun::set_envvar(c(RMARKDOWN_LUA_SHARED = pkg_file_lua("shared.lua")))
+    on.exit(xfun::set_envvar(lua_env_vars), add = TRUE)
+
     perf_timer_start("pre-processor")
 
     # call any pre_processor
@@ -889,8 +893,6 @@ render <- function(input,
 
       # if Lua filters are provided, add the command line switch
       if (!is.null(lua_filters <- output_format$pandoc$lua_filters)) {
-        lua_env_vars <- xfun::set_envvar(c(RMARKDOWN_LUA_SHARED = pkg_file_lua("shared.lua")))
-        on.exit(xfun::set_envvar(lua_env_vars), add = TRUE)
         lua_filters <- pandoc_lua_filter_args(lua_filters)
       }
       pandoc_args <- c(lua_filters, pandoc_args)