+++ R/slidy_presentation.R
@@ -46,6 +46,7 @@ slidy_presentation <- function(number_sections = FALSE,
                                df_print = "default",
                                self_contained = TRUE,
                                highlight = "default",
+                               math_method = "default",
                                mathjax = "default",
                                template = "default",
                                css = NULL,
@@ -119,6 +120,7 @@ slidy_presentation <- function(number_sections = FALSE,
     args <- c()
 
     # highlight
+    if (!is.null(highlight)) highlight <- resolve_highlight(highlight, highlighters())
     args <- c(args, pandoc_highlight_args(highlight, default = "pygments"))
 
     # return additional args
@@ -140,6 +142,7 @@ slidy_presentation <- function(number_sections = FALSE,
     pre_processor = pre_processor,
     base_format = html_document_base(lib_dir = lib_dir,
                                      self_contained = self_contained,
+                                     math_method = math_method,
                                      mathjax = mathjax,
                                      bootstrap_compatible = TRUE,
                                      pandoc_args = pandoc_args,